Update example - create-react-native app
