+++ .eslintrc
@@ -5,7 +5,8 @@
     "react",
     "flowtype",
     "jsx-a11y",
-    "import"
+    "import",
+    "babel"
   ],
   "rules": {
     // Best Practices
@@ -42,7 +43,7 @@
     "no-floating-decimal": 2,
     "no-implicit-coercion": 2,
     "no-implied-eval": 2,
-    "no-invalid-this": 2,
+    "no-invalid-this": 0,
     "no-iterator": 2,
     "no-labels": 2,
     "no-lone-blocks": 2,
