+++ Example/.babelrc
@@ -1,3 +1,12 @@
 {
-  "presets": ["react-native"]
+  "presets": [
+    "babel-preset-expo"
+  ],
+  "env": {
+    "development": {
+      "plugins": [
+        "transform-react-jsx-source"
+      ]
+    }
+  }
 }
