+++ Example/.watchmanconfig
