--- Example/android/settings.gradle
@@ -1,3 +1,0 @@
-rootProject.name = 'Example'
-
-include ':app'
