--- Example/index.ios.js
@@ -1,16 +1,0 @@
-/**
- * Sample React Native App
- * https://github.com/facebook/react-native
- * @flow
- */
-
-import React from 'react'
-import {
-  AppRegistry,
-} from 'react-native'
-
-import ExampleMain from './ExampleMain'
-
-const Example = () => <ExampleMain />
-
-AppRegistry.registerComponent('Example', () => Example)
