+++ example/App.js
@@ -1,0 +1,10 @@
+/* @flow */
+
+import React from 'react'
+import SegmentedView from './SegmentedView'
+
+export default function App() {
+  return (
+    <SegmentedView />
+  )
+}
