+++ package.json
@@ -34,6 +34,7 @@
     "babel-preset-react-native": "^4.0.1",
     "eslint": "^5.7.0",
     "eslint-config-airbnb": "^17.1.0",
+    "eslint-plugin-babel": "^5.2.1",
     "eslint-plugin-flowtype": "^3.0.0",
     "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jsx-a11y": "^6.1.2",
